+++ packages/chatbot-server-mongodb-public/src/config.ts
@@ -16,11 +16,11 @@ import {
   requireValidIpAddress,
   requireRequestOrigin,
   AddCustomDataFunc,
-  makeVerifiedAnswerGenerateUserPrompt,
   makeDefaultFindVerifiedAnswer,
   defaultCreateConversationCustomData,
   defaultAddMessageToConversationCustomData,
   makeLegacyGeneratateResponse,
+  makeVerifiedAnswerGenerateResponse,
 } from "mongodb-chatbot-server";
 import cookieParser from "cookie-parser";
 import { makeStepBackRagGenerateUserPrompt } from "./processors/makeStepBackRagGenerateUserPrompt";
@@ -174,8 +174,8 @@ export const preprocessorOpenAiClient = wrapOpenAI(
   })
 );
 
-export const generateUserPrompt = wrapTraced(
-  makeVerifiedAnswerGenerateUserPrompt({
+export const generateResponse = wrapTraced(
+  makeVerifiedAnswerGenerateResponse({
     findVerifiedAnswer,
     onVerifiedAnswerFound: (verifiedAnswer) => {
       return {
@@ -184,11 +184,17 @@ export const generateUserPrompt = wrapTraced(
       };
     },
     onNoVerifiedAnswerFound: wrapTraced(
-      makeStepBackRagGenerateUserPrompt({
-        openAiClient: preprocessorOpenAiClient,
-        model: retrievalConfig.preprocessorLlm,
-        findContent,
-        numPrecedingMessagesToInclude: 6,
+      makeLegacyGeneratateResponse({
+        llm,
+        generateUserPrompt: makeStepBackRagGenerateUserPrompt({
+          openAiClient: preprocessorOpenAiClient,
+          model: retrievalConfig.preprocessorLlm,
+          findContent,
+          numPrecedingMessagesToInclude: 6,
+        }),
+        systemMessage: systemPrompt,
+        llmNotWorkingMessage: "LLM not working. Sad!",
+        noRelevantContentMessage: "No relevant content found. Sad!",
       }),
       { name: "makeStepBackRagGenerateUserPrompt" }
     ),
@@ -294,13 +300,7 @@ export const config: AppConfig = {
           : undefined,
       segment: segmentConfig,
     }),
-    generateResponse: makeLegacyGeneratateResponse({
-      llm,
-      generateUserPrompt,
-      systemMessage: systemPrompt,
-      llmNotWorkingMessage: "LLM not working. Sad!",
-      noRelevantContentMessage: "No relevant content found. Sad!",
-    }),
+    generateResponse,
     maxUserMessagesInConversation: 50,
     maxUserCommentLength: 500,
     conversations,
